Refactor recorder architecture and improve session handling
