+++ Test_App/URL+Identifiable.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+extension URL: @retroactive Identifiable {
+    public var id: URL { self }
+}
